requirements.txt
selenium>=4.10.0
pandas>=1.3.0
openai>=1.0.0
python-dotenv>=0.19.0